requirements.txt
torch==2.4.1
torchvision==0.19.1
./models/medclip/MedCLIP
numpy
pandas
Pillow
requests
tqdm
wget
nltk>=3.7
scikit_learn>=1.1.2
textaugment>=1.3.4
timm==1.0.9
transformers==4.24.0
mlflow==2.14.3
cffi==1.17.1
cloudpickle==3.0.0
colorama==0.4.6
einops==0.8.0
ftfy==6.2.3
fvcore==0.1.5.post20221221
mup==1.0.0
packaging==24.1
pyyaml==6.0.2
sentencepiece==0.2.0
tenacity==9.0.0
tornado==6.4.1
open_clip_torch
matplotlib
hi-ml-multimodal
pydicom==2.1.2
omegaconf==2.3.0
tensorboard==2.13.0
albumentations==1.3.1